+++ src/main/java/lotto/view/InputView.java
@@ -3,25 +3,26 @@ package lotto.view;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import lotto.util.InputValidator;
+import java.util.List;
 
 public class InputView {
 
+    public static final String DELIMITER = ",";
     private static final String LINE = System.lineSeparator();
 
-    public String readPurchasePrice() {
+    public int readPurchasePrice() {
         System.out.println("구입금액을 입력해 주세요.");
-        return getInput();
+        return StringToIntParser.parseInt(getInput());
     }
 
-    public String readWinningLottoNumbers() {
+    public List<Integer> readWinningLottoNumbers() {
         System.out.println(LINE + "지난 주 당첨 번호를 입력해 주세요.");
-        return getInput();
+        return StringToIntParser.parseTokens(getInput().split(DELIMITER));
     }
 
-    public String readBonusNumber() {
+    public int readBonusNumber() {
         System.out.println("보너스 볼을 입력해 주세요.");
-        return getInput();
+        return StringToIntParser.parseInt(getInput());
     }
 
     private String getInput() {
