+++ src/main/java/lotto/view/StringToIntParser.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.List;
