+++ src/test/java/lotto/view/InputValidatorTest.java
@@ -1,5 +1,6 @@
-package lotto.util;
+package lotto.view;
 
+import lotto.view.InputValidator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
