+++ src/test/java/lotto/view/StringToIntParserTest.java
@@ -1,6 +1,8 @@
-package lotto.util;
+package lotto.view;
 
 import java.util.List;
+
+import lotto.view.StringToIntParser;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
